# api/openapi.yaml
openapi: 3.0.3
info:
  title: BagiCore
  description: Go-based comic hosting catalog backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
servers:
  - url: /api/v0
paths:
...
  /comics/{code}/links:
    post:
      tags:
        - Comic
      summary: Add comic link.
      operationId: addComicLink
      parameters:
        - name: code
          in: path
          description: Code of comic.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewComicLink'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewComicLink'
        required: true
      responses:
        '201':
          description: Comic link added.
          headers:
            Location:
              description: The path of new comic link.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicLink'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
...
  /comics/{code}/chapters/{cv}/links:
    post:
      tags:
        - Comic
      summary: Add comic chapter link.
      operationId: addComicChapterLink
      parameters:
        - name: code
          in: path
          description: Code of comic.
          required: true
          schema:
            type: string
        - name: cv
          in: path
          description: Chapter[+Version] of comic chapter.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewComicChapterLink'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewComicChapterLink'
        required: true
      responses:
        '201':
          description: Comic chapter link added.
          headers:
            Location:
              description: The path of new comic chapter link.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComicChapterLink'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
...
  /links:
    get:
      tags:
        - Link
      summary: List link.
      operationId: listLink
      parameters:
        - name: page
          in: query
          description: Page number of results.
          schema:
            type: integer
        - name: limit
          in: query
          description: Maximum number of results.
          schema:
            type: integer
        - name: order_by
          in: query
          description: Sort results returned.
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Link list.
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: The total count of link with current filter.
            X-Pagination-Limit:
              schema:
                type: integer
              description: The last page number of link with current filter and limit.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Link'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Link
      summary: Add link.
      operationId: addLink
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewLink'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewLink'
        required: true
      responses:
        '201':
          description: Link added.
          headers:
            Location:
              description: The path of new link.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /links/{websiteDomain}-{relativeURL}:
    get:
      tags:
        - Link
      summary: Get link.
      operationId: getLink
      parameters:
        - name: websiteDomain
          in: path
          description: Website domain name of link to return.
          required: true
          schema:
            type: string
        - name: relativeURL
          in: path
          description: Relative URL of link to return.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Link gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Link
      summary: Update link.
      operationId: updateLink
      parameters:
        - name: websiteDomain
          in: path
          description: Website domain name of link to update.
          required: true
          schema:
            type: string
        - name: relativeURL
          in: path
          description: Relative URL of link to update.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetLink'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SetLink'
        required: true
      responses:
        '200':
          description: Link updated.
          headers:
            Location:
              description: The path of updated link.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        '204':
          description: Link unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Link
      summary: Delete link.
      operationId: deleteLink
      parameters:
        - name: websiteDomain
          in: path
          description: Website domain name of linik to delete.
          required: true
          schema:
            type: string
        - name: relativeURL
          in: path
          description: Relative URL of linik to delete.
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Link deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /links/{websiteDomain}-{relativeURL}/tl-languages:
    post:
      tags:
        - Link
      summary: Add link TL language.
      operationId: addLinkTLLanguage
      parameters:
        - name: websiteDomain
          in: path
          description: Website domain name of link.
          required: true
          schema:
            type: string
        - name: relativeURL
          in: path
          description: Relative URL of link.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewLinkTLLanguage'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewLinkTLLanguage'
        required: true
      responses:
        '201':
          description: Link TL language added.
          headers:
            Location:
              description: The path of new link TL language.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkTLLanguage'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /links/{websiteDomain}-{relativeURL}/tl-languages/{ietf}:
    get:
      tags:
        - Link
      summary: Get link TL language.
      operationId: getLinkTLLanguage
      parameters:
        - name: websiteDomain
          in: path
          description: Website domain name of link.
          required: true
          schema:
            type: string
        - name: relativeURL
          in: path
          description: Relative URL of link.
          required: true
          schema:
            type: string
        - name: ietf
          in: path
          description: IETF code of link TL language to return.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Link TL language gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkTLLanguage'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Link
      summary: Update link TL language.
      operationId: updateLinkTLLanguage
      parameters:
        - name: websiteDomain
          in: path
          description: Website domain name of link.
          required: true
          schema:
            type: string
        - name: relativeURL
          in: path
          description: Relative URL of link.
          required: true
          schema:
            type: string
        - name: ietf
          in: path
          description: IETF code of link TL language to update.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetLinkTLLanguage'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SetLinkTLLanguage'
        required: true
      responses:
        '200':
          description: Link TL language updated.
          headers:
            Location:
              description: The path of updated link TL language.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkTLLanguage'
        '204':
          description: Link TL language unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Link
      summary: Delete link TL language.
      operationId: deleteLinkTLLanguage
      parameters:
        - name: websiteDomain
          in: path
          description: Website domain name of link.
          required: true
          schema:
            type: string
        - name: relativeURL
          in: path
          description: Relative URL of link.
          required: true
          schema:
            type: string
        - name: ietf
          in: path
          description: IETF code of link TL language to delete.
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Link TL language deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
components:
  schemas:
    Object:
      type: object
      properties:
        id:
          type: integer
          format: int64
          x-go-type: uint
          x-go-name: ID
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - createdAt
...
    ComicLink:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        linkID:
          type: integer
          x-go-type: uint
        linkWebsiteDomain:
          type: string
        linkRelativeURL:
          type: string
      required:
        - createdAt
        - linkID
        - linkWebsiteDomain
        - linkRelativeURL
    NewComicLink:
      type: object
      properties:
        linkID:
          type: integer
          nullable: true
          x-go-type: uint
        linkWebsiteDomain:
          type: string
          nullable: true
        linkRelativeURL:
          type: string
          nullable: true
...
    ComicChapterLink:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        linkID:
          type: integer
          x-go-type: uint
        linkWebsiteDomain:
          type: string
        linkRelativeURL:
          type: string
      required:
        - createdAt
        - linkID
        - linkWebsiteDomain
        - linkRelativeURL
    NewComicChapterLink:
      type: object
      properties:
        linkID:
          type: integer
          nullable: true
          x-go-type: uint
        linkWebsiteDomain:
          type: string
          nullable: true
        linkRelativeURL:
          type: string
          nullable: true
...
    Language:
      type: object
      allOf:
        - $ref: '#/components/schemas/Object'
        - type: object
          properties:
            ietf:
              type: string
              x-go-name: IETF
            name:
              type: string
          required: 
            - ietf
            - name
...
    Link:
      type: object
      allOf:
        - $ref: '#/components/schemas/Object'
        - type: object
          properties:
            websiteID:
              type: integer
              x-go-type: uint
            websiteDomain:
              type: string
            relativeURL:
              type: string
            tlLanguages:
              type: array
              items:
                $ref: '#/components/schemas/Language'
              x-go-name: TLLanguages
            machineTL:
              type: boolean
              nullable: true
          required: 
            - websiteID
            - websiteDomain
            - relativeURL
    NewLink:
      type: object
      properties:
        websiteID:
          type: integer
          nullable: true
          x-go-type: uint
        websiteDomain:
          type: string
          nullable: true
        relativeURL:
          type: string
        machineTL:
          type: boolean
          nullable: true
      required: 
        - relativeURL
    SetLink:
      type: object
      properties:
        websiteID:
          type: integer
          nullable: true
          x-go-type: uint
        websiteDomain:
          type: string
          nullable: true
        relativeURL:
          type: string
          nullable: true
        machineTL:
          type: boolean
          nullable: true
        setNull:
          type: array
          items:
            type: string
          x-go-type-skip-optional-pointer: true
    LinkTLLanguage:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        languageID:
          type: integer
          x-go-type: uint
        languageIETF:
          type: string
      required:
        - createdAt
        - languageID
        - languageIETF
    NewLinkTLLanguage:
      type: object
      properties:
        languageID:
          type: integer
          nullable: true
          x-go-type: uint
        languageIETF:
          type: string
          nullable: true
    SetLinkTLLanguage:
      type: object
      properties:
        languageID:
          type: integer
          nullable: true
          x-go-type: uint
        languageIETF:
          type: string
          nullable: true
    Error:
      type: object
      properties:
        error:
          type: object
          properties: 
            status:
              type: string
            message:
              type: string
          required:
            - status
            - message
      required:
        - error
  responses:
    Default:
      description: Unexpected error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT